Rename Fast API serializers/schemas to `datamodels`
Pierre, Ash & I discussed about making the naming consistent for Pydantic Models. We agreed on `datamodels` since it is what OpenAPI terms it too apart from schemas: https://swagger.io/docs/specification/v3_0/data-models/data-models/.

We didn't choose `models` because we have DB models named as such.
