+++ airflow/api_fastapi/core_api/datamodels/plugins.py
