+++ airflow/api_fastapi/core_api/datamodels/trigger.py
