+++ airflow/api_fastapi/core_api/openapi/v1-generated.yaml
@@ -4336,7 +4336,7 @@ components:
         dag_run_states:
           $ref: '#/components/schemas/DAGRunStates'
         task_instance_states:
-          $ref: '#/components/schemas/airflow__api_fastapi__core_api__serializers__ui__dashboard__TaskInstanceState'
+          $ref: '#/components/schemas/airflow__api_fastapi__core_api__datamodels__ui__dashboard__TaskInstanceState'
       type: object
       required:
       - dag_run_types
@@ -5056,7 +5056,7 @@ components:
       - git_version
       title: VersionInfo
       description: Version information serializer for responses.
-    airflow__api_fastapi__core_api__serializers__ui__dashboard__TaskInstanceState:
+    airflow__api_fastapi__core_api__datamodels__ui__dashboard__TaskInstanceState:
       properties:
         no_status:
           type: integer
