+++ airflow/api_fastapi/core_api/routes/public/event_logs.py
@@ -33,11 +33,11 @@ from airflow.api_fastapi.common.parameters import (
     SortParam,
 )
 from airflow.api_fastapi.common.router import AirflowRouter
-from airflow.api_fastapi.core_api.openapi.exceptions import create_openapi_http_exception_doc
-from airflow.api_fastapi.core_api.serializers.event_logs import (
+from airflow.api_fastapi.core_api.datamodels.event_logs import (
     EventLogCollectionResponse,
     EventLogResponse,
 )
+from airflow.api_fastapi.core_api.openapi.exceptions import create_openapi_http_exception_doc
 from airflow.models import Log
 
 event_logs_router = AirflowRouter(tags=["Event Log"], prefix="/eventLogs")
