+++ airflow/api_fastapi/core_api/routes/public/monitor.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from airflow.api.common.airflow_health import get_airflow_health
 from airflow.api_fastapi.common.router import AirflowRouter
-from airflow.api_fastapi.core_api.serializers.monitor import HealthInfoSchema
+from airflow.api_fastapi.core_api.datamodels.monitor import HealthInfoSchema
 
 monitor_router = AirflowRouter(tags=["Monitor"], prefix="/monitor")
