+++ airflow/api_fastapi/core_api/routes/public/plugins.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from airflow.api_fastapi.common.parameters import QueryLimit, QueryOffset
 from airflow.api_fastapi.common.router import AirflowRouter
-from airflow.api_fastapi.core_api.serializers.plugins import PluginCollectionResponse, PluginResponse
+from airflow.api_fastapi.core_api.datamodels.plugins import PluginCollectionResponse, PluginResponse
 from airflow.plugins_manager import get_plugin_info
 
 plugins_router = AirflowRouter(tags=["Plugin"], prefix="/plugins")
